backend/requirements.txt
Django==4.2.4
djangorestframework==3.14.0
psycopg2-binary==2.9.8
gunicorn==21.2.0
pillow==10.0.1
drf-spectacular==0.27.2